Added movement class for movement with velocity

+++ src/references.cpp
@@ -6,7 +6,7 @@
 #include "scene/camera.hpp"
 
 int References::fps = 0;
-double References::frame_time = 0;
+float References::frame_time = 0;
 int References::width = 1600;
 int References::height = 900;
 Camera* References::main_camera = NULL;
